Change schemas structure
