+++ db/seed_db.py
@@ -5,7 +5,7 @@ from .database import (
     Base,
 )
 from .models import DbCategory
-from schemas.mov_dir_actors_schemas import MovieCategoryType
+from schemas.categories_schemas import MovieCategoryType
 
 
 def seed_categories(session: Session):
